# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: TravelPlan API
  version: 1.0.0
  description: >-
    # TravelPlan Public API


    Production-grade AI travel management API.


    ## Versioning


    - **Current:** `/api/v1`

    - **Legacy (compatible):** `/api` (same handlers; receives `X-API-Version:
    1`)


    Prefer `/api/v1` for all new integrations.


    ## Authentication


    Most endpoints require a JWT Bearer token from `POST /api/v1/auth/login` or
    signup.


    ```

    Authorization: Bearer <token>

    ```


    Admin monitoring APIs additionally require an admin role (`User.role=admin`
    or `ADMIN_EMAILS`).


    ## Rate limits


    Redis-backed limits (fail-open if Redis is down):


    | Scope | Limit |

    |-------|-------|

    | Login | 10 / minute |

    | Signup | 10 / minute |

    | Forgot password | 5 / minute |

    | AI endpoints | 20 / hour |

    | Public weather | 60 / minute |


    Responses may include `X-RateLimit-*` headers. HTTP **429** when exceeded.


    ## Errors


    Error bodies use:


    ```json

    { "success": false, "code": "ERROR_CODE", "message": "Human readable",
    "error": "..." }

    ```
  contact:
    name: TravelPlan API
    email: hello@travelplan.com
  license:
    name: Proprietary
    identifier: LicenseRef-TravelPlan-Proprietary
servers:
  - url: http://localhost:5000
    description: Local development (paths include /api/v1)
  - url: https://api.travelplan.example
    description: Production host (example)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT from /auth/login or /auth/signup. Send as `Authorization:
        Bearer <token>`."
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data: {}
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
        error:
          type: string
...
paths:
...
  /api/v1/recommendations:
    get:
      summary: Recommendations
      description: Recommendations
      tags: &a187
        - Recommendations
      parameters: &a188 []
      responses: &a189
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: getRecommendations
      security: &a190 []
  /api/recommendations:
    get:
      summary: Recommendations
      description: |-
        Recommendations

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a187
      parameters: *a188
      responses: *a189
      operationId: getRecommendations
      security: *a190
      x-deprecated-prefer: Use /api/v1getRecommendations
...
  /api/v1/activities:
    get:
      summary: Search activities
      description: Search activities
      tags: &a331
        - Availability
      parameters: &a332
        - name: destination
          in: query
          required: false
          description: Destination
          schema:
            type: string
        - name: date
          in: query
          required: false
          description: Date YYYY-MM-DD
          schema:
            type: string
      responses: &a333
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      operationId: searchActivities
      security: &a334 []
  /api/activities:
    get:
      summary: Search activities
      description: |-
        Search activities

        **Legacy alias** of the `/api/v1` path. Prefer versioned URLs.
      tags: *a331
      parameters: *a332
      responses: *a333
      operationId: searchActivities
      security: *a334
      x-deprecated-prefer: Use /api/v1searchActivities
...